fix!: remove default stream close event (#81)
* fix!: remove default stream close event

BREAKING CHANGE: there is no "end" event when stream is closing in case you depended on it

The current implementation sends its own event when an async iterable ends, this is bad as the implemented SSE  protocol doesn't expect that message, it's an application level thing. It should be up to the application to send such a message or just close the SSE stream.

Remove that event (Possibly a breaking change if anyone relied on it).

For example, that caused an error in our project from our handler receving that message and crashing cause it's an unknown message type whose body is not even JSON as our application expects.

Contributed on behalf of [Swimm](https://swimm.io/)

* fix tests

Signed-off-by: Marin Petrunic <marin.petrunic@gmail.com>

* update ci nodejs version

Signed-off-by: Marin Petrunic <marin.petrunic@gmail.com>

---------

Signed-off-by: Marin Petrunic <marin.petrunic@gmail.com>
Co-authored-by: Marin Petrunic <marin.petrunic@gmail.com>